Generate header id to work with gitbook 3.0.0
gitbook 3.0.0 does not seem to generate the `id` attribute for headers
before it sends the content to plugins. So, the call to
`header.attr('id')` returned `undefined`.

This commit fixes it by generating an `id`. The slug is generated with
`github-slugid` module.

For backwards compatibility, if an existing `id` is present it is used
instead of creating an id.

Signed-off-by: Kaustav Das Modak <kaustav.dasmodak@yahoo.co.in>